+++ ChangeLog.md
@@ -3,6 +3,10 @@
 ## 0.7.2
 
 - Add 'ObserverState' type alias
+- Add 'ObservationQueue' utility
+- Fix missing re-exports from
+  `Control.Eff.Concurrent.Api.Client`
+  in `Control.Eff.Concurrent` introduced in recent versions
 
 ## 0.7.1